+++ gazelle/python/language.go
@@ -23,7 +23,6 @@ import (
 type Python struct {
 	Configurer
 	Resolver
-	LifeCycleManager
 }
 
 // NewLanguage initializes a new Python that satisfies the language.Language