+++ CHANGELOG.md
@@ -22,6 +22,8 @@ A brief description of the categories of changes:
 [x.x.x]: https://github.com/bazelbuild/rules_python/releases/tag/x.x.x
 
 ### Changed
+* **BREAKING** (gazelle): Remove gazelle plugin's python deps and make it hermetic. Use a new helper
+  based on tree-sitter and written in go for syntax analysis.
 
 ### Fixed